Rename svg generator file and fix class usage
