+++ svg_generator.rb
